// src/app/chatbot/models/interaction.ts
export interface Interaction {
    name?: string;
    slug?: string;
    question_type?: string;
    message_type?: string;
    action_type?: string;
    interaction_type: 'question' | 'message' | 'action';
    interaction_text?: string;
    selected?: boolean;
    errSelected?: boolean;
    scheduleSetupPending?: boolean;
    counter?: number;
    number_low?: number;
    number_high?: number;
    answer_character_limit?: number;
    time_start?: string;
    time_end?: string;
    require_in_future?: boolean;
    url?: string;
    title?: string;
    multi_choice_minimum?: number;
    multi_choice_maximum?: number;
    image?: string;
    document?: string;
    video?: string;
    interaction_options?: any[];
    multi_choice_type?: string;
    media_status?: string;
    file_key?: string;
    imageValue?: string;
    videoValue?: string;
    documentValue?: string;
    copiedStatus?: boolean;
    ask_consent_frequency?: string;
    saving_to_both?: boolean;
    contact_field_id?: string;
    contact_field_type?: number;
    ask_if_exists?: boolean;
    deleteMultiChoiceOptions?: boolean;
    schedule_ask_detail_message?: string;
    schedule_first_name?: string;
    schedule_last_name?: string;
    schedule_email?: string;
    schedule_fallback_message?: string;
    schedule_timeslot_fallback_message?: string;
    schedule_additional_settings?: boolean;
    schedule_timeslot_fallback_button_text?: string;
    schedule_action_event_type_meeting_length?: number;
    schedule_action_time_zone_conversion?: boolean;
    schedule_action_time_zone_conversion_prompt?: boolean;
    schedule_action_booking_page_time_zone?: number;
    button_text?: string;
    schedule_master_booking_page_id?: string;
    schedule_event_type?: string;
    schedule_event_type_external_id?: string;
    schedule_action_event_type?: string;
    schedule_panel_id?: string;
    schedule_master_page_image?: string;
    schedule_confirmation_title?: string;
    schedule_confirmation_button?: string;
    routingFullName?: string;
    is_welcome_interaction?: boolean;
    otherwise_route?: any;
    confirm_consent?: boolean;
    bot_response?: string;
    basic_route_slug?: string;
    label?: string;
    icon?: string;
    schedule_mobile_phone_text?: string;
    schedule_sms_opt_in_text?: string;
    schedule_sms_notification?: boolean;
    schedule_action_panel_members?: any;
    sub_title?: string;
    audio?: string;
    cards?: any[];
    jump_to_interaction?: any;
    match_send_type?: string;
    timeout_period?: number;
    handoff_message?: string;
    away_message?: string;
    email_alert_subject?: string;
    email_alert_message_body?: string;
    icon_type?: string;
    is_form_title?: boolean;
    show_form_title?: boolean;
    email_alert_recipients?: any[];
  }
  